ezid and batch-register are now installed properly. ezid handles different servers.
