+++ ezid_client_tools/batch_register.py
