+++ ezid_client_tools/ezid.py
@@ -6,7 +6,7 @@
 # default, ANVL responses (currently, that's all responses) are left
 # in %-encoded form.
 #
-# Usage: ezid.py [options] credentials operation...
+# Usage: ezid [options] credentials operation...
 #
 #   options:
 #     -d          decode ANVL responses
@@ -64,11 +64,13 @@ import re
 import sys
 import time
 import types
-import urllib
-import urllib2
+import urllib.request as urlreq # Python 3.x
+from urllib.parse import quote # Python 3.x
 
 KNOWN_SERVERS = {
-  "p": "https://ezid.cdlib.org"
+    "l": "http://localhost:8000",
+    "s": "https://uc3-ezidx2-stg.cdlib.org",
+    "p": "https://ezid.cdlib.org",
 }
 
 OPERATIONS = {
@@ -83,13 +85,22 @@ OPERATIONS = {
   "status": (0, False)
 }
 
-USAGE_TEXT = """Usage: ezid.py [options] credentials operation...
+USAGE_TEXT = """Usage: ezid [options] credentials operation...
 
   options:
     -d          decode ANVL responses
     -e ENCODING character encoding; defaults to UTF-8
     -o          one line per ANVL value: convert newlines to spaces
     -t          format timestamps
+
+  Positional arguments:
+
+  server:
+    l (local server, no https)
+    s (staging)
+    p (production)
+    http[s]://...
+
 
   credentials:
     username:password
@@ -121,19 +132,19 @@ class MyHelpFormatter (optparse.IndentedHelpFormatter):
   def format_usage (self, usage):
     return USAGE_TEXT
 
-class MyHTTPErrorProcessor (urllib2.HTTPErrorProcessor):
+class MyHTTPErrorProcessor (urlreq.HTTPErrorProcessor):
   def http_response (self, request, response):
     # Bizarre that Python leaves this out.
     if response.code == 201:
       return response
     else:
-      return urllib2.HTTPErrorProcessor.http_response(self, request, response)
+      return urlreq.HTTPErrorProcessor.http_response(self, request, response)
   https_response = http_response
 
 def formatAnvlRequest (args):
   request = []
   for i in range(0, len(args), 2):
-    k = args[i].decode(_options.encoding)
+    k = args[i]
     if k == "@":
       f = codecs.open(args[i+1], encoding=_options.encoding)
       request += [l.strip("\r\n") for l in f.readlines()]
@@ -143,7 +154,7 @@ def formatAnvlRequest (args):
         k = "@"
       else:
         k = re.sub("[%:\r\n]", lambda c: "%%%02X" % ord(c.group(0)), k)
-      v = args[i+1].decode(_options.encoding)
+      v = args[i+1]
       if v.startswith("@@"):
         v = v[1:]
       elif v.startswith("@") and len(v) > 1:
@@ -155,15 +166,15 @@ def formatAnvlRequest (args):
   return "\n".join(request)
 
 def encode (id):
-  return urllib.quote(id, ":/")
+  return quote(id, ":/")
 
 def issueRequest (path, method, data=None, returnHeaders=False,
   streamOutput=False):
-  request = urllib2.Request("%s/%s" % (_server, path))
+  request = urlreq.Request("%s/%s" % (_server, path))
   request.get_method = lambda: method
   if data:
     request.add_header("Content-Type", "text/plain; charset=UTF-8")
-    request.add_data(data.encode("UTF-8"))
+    request.data = data.encode("UTF-8")
   if _cookie: request.add_header("Cookie", _cookie)
   try:
     connection = _opener.open(request)
@@ -177,7 +188,7 @@ def issueRequest (path, method, data=None, returnHeaders=False,
         return response.decode("UTF-8"), connection.info()
       else:
         return response.decode("UTF-8")
-  except urllib2.HTTPError, e:
+  except urlreq.HTTPError as e:
     sys.stderr.write("%d %s\n" % (e.code, e.msg))
     if e.fp != None:
       response = e.fp.read()
@@ -202,103 +213,111 @@ def printAnvlResponse (response, sortLines=False):
       line = re.sub("%([0-9a-fA-F][0-9a-fA-F])",
         lambda m: chr(int(m.group(1), 16)), line)
     if _options.oneLine: line = line.replace("\n", " ").replace("\r", " ")
-    print line.encode(_options.encoding)
+    print(line.encode(_options.encoding))
 
 # Process command line arguments.
 
-parser = optparse.OptionParser(formatter=MyHelpFormatter())
-parser.add_option("-d", action="store_true", dest="decode", default=False)
-parser.add_option("-e", action="store", dest="encoding", default="UTF-8")
-parser.add_option("-o", action="store_true", dest="oneLine", default=False)
-parser.add_option("-t", action="store_true", dest="formatTimestamps",
-  default=False)
+def main():
 
-_options, args = parser.parse_args()
-# Simulate selection of the production server (server selection is not
-# supported in this public version of the code).
-args.insert(0, "p")
-if len(args) < 3: parser.error("insufficient arguments")
+  global _options, _server, _opener, _cookie
 
-_server = KNOWN_SERVERS.get(args[0], args[0])
+  parser = optparse.OptionParser(formatter=MyHelpFormatter())
+  parser.add_option("-d", action="store_true", dest="decode", default=False)
+  parser.add_option("-e", action="store", dest="encoding", default="UTF-8")
+  parser.add_option("-o", action="store_true", dest="oneLine", default=False)
+  parser.add_option("-t", action="store_true", dest="formatTimestamps",
+    default=False)
 
-_opener = urllib2.build_opener(MyHTTPErrorProcessor())
-if args[1].startswith("sessionid="):
-  _cookie = args[1]
-elif args[1] != "-":
-  if ":" in args[1]:
-    username, password = args[1].split(":", 1)
+  _options, args = parser.parse_args()
+  # Simulate selection of the production server (server selection is not
+  # supported in this public version of the code).
+  # args.insert(0, "p")
+
+  if len(args) < 3: parser.error("insufficient arguments")
+
+  _server = KNOWN_SERVERS.get(args[0], args[0])
+
+  _opener = urlreq.build_opener(MyHTTPErrorProcessor())
+  if args[1].startswith("sessionid="):
+    _cookie = args[1]
+  elif args[1] != "-":
+    if ":" in args[1]:
+      username, password = args[1].split(":", 1)
+    else:
+      username = args[1]
+      password = getpass.getpass()
+    h = urlreq.HTTPBasicAuthHandler()
+    h.add_password("EZID", _server, username, password)
+    _opener.add_handler(h)
+
+  if args[2].endswith("!"):
+    bang = True
+    args[2] = args[2][:-1]
   else:
-    username = args[1]
-    password = getpass.getpass()
-  h = urllib2.HTTPBasicAuthHandler()
-  h.add_password("EZID", _server, username, password)
-  _opener.add_handler(h)
+    bang = False
+  operation = list(filter(lambda o: o.startswith(args[2]), OPERATIONS))
+  if len(operation) != 1: parser.error("unrecognized or ambiguous operation")
+  operation = operation[0]
+  if bang and not OPERATIONS[operation][1]:
+    parser.error("unrecognized operation")
 
-if args[2].endswith("!"):
-  bang = True
-  args[2] = args[2][:-1]
-else:
-  bang = False
-operation = filter(lambda o: o.startswith(args[2]), OPERATIONS)
-if len(operation) != 1: parser.error("unrecognized or ambiguous operation")
-operation = operation[0]
-if bang and not OPERATIONS[operation][1]:
-  parser.error("unrecognized operation")
+  args = args[3:]
 
-args = args[3:]
+  if (type(OPERATIONS[operation][0]) is int and\
+    len(args) != OPERATIONS[operation][0]) or\
+    (type(OPERATIONS[operation][0]) is types.LambdaType and\
+    not OPERATIONS[operation][0](len(args))):
+    parser.error("incorrect number of arguments for operation")
 
-if (type(OPERATIONS[operation][0]) is int and\
-  len(args) != OPERATIONS[operation][0]) or\
-  (type(OPERATIONS[operation][0]) is types.LambdaType and\
-  not OPERATIONS[operation][0](len(args))):
-  parser.error("incorrect number of arguments for operation")
+  # Perform the operation.
 
-# Perform the operation.
+  if operation == "mint":
+    shoulder = args[0]
+    if len(args) > 1:
+      data = formatAnvlRequest(args[1:])
+    else:
+      data = None
+    response = issueRequest("shoulder/"+encode(shoulder), "POST", data)
+    printAnvlResponse(response)
+  elif operation == "create":
+    id = args[0]
+    if len(args) > 1:
+      data = formatAnvlRequest(args[1:])
+    else:
+      data = None
+    path = "id/"+encode(id)
+    if bang: path += "?update_if_exists=yes"
+    response = issueRequest(path, "PUT", data)
+    printAnvlResponse(response)
+  elif operation == "view":
+    id = args[0]
+    path = "id/"+encode(id)
+    if bang: path += "?prefix_match=yes"
+    response = issueRequest(path, "GET")
+    printAnvlResponse(response, sortLines=True)
+  elif operation == "update":
+    id = args[0]
+    if len(args) > 1:
+      data = formatAnvlRequest(args[1:])
+    else:
+      data = None
+    response = issueRequest("id/"+encode(id), "POST", data)
+    printAnvlResponse(response)
+  elif operation == "delete":
+    id = args[0]
+    response = issueRequest("id/"+encode(id), "DELETE")
+    printAnvlResponse(response)
+  elif operation == "login":
+    response, headers = issueRequest("login", "GET", returnHeaders=True)
+    response += "\nsessionid=%s\n" %\
+      headers["set-cookie"].split(";")[0].split("=")[1]
+    printAnvlResponse(response)
+  elif operation == "logout":
+    response = issueRequest("logout", "GET")
+    printAnvlResponse(response)
+  elif operation == "status":
+    response = issueRequest("status", "GET")
+    printAnvlResponse(response)
 
-if operation == "mint":
-  shoulder = args[0]
-  if len(args) > 1:
-    data = formatAnvlRequest(args[1:])
-  else:
-    data = None
-  response = issueRequest("shoulder/"+encode(shoulder), "POST", data)
-  printAnvlResponse(response)
-elif operation == "create":
-  id = args[0]
-  if len(args) > 1:
-    data = formatAnvlRequest(args[1:])
-  else:
-    data = None
-  path = "id/"+encode(id)
-  if bang: path += "?update_if_exists=yes"
-  response = issueRequest(path, "PUT", data)
-  printAnvlResponse(response)
-elif operation == "view":
-  id = args[0]
-  path = "id/"+encode(id)
-  if bang: path += "?prefix_match=yes"
-  response = issueRequest(path, "GET")
-  printAnvlResponse(response, sortLines=True)
-elif operation == "update":
-  id = args[0]
-  if len(args) > 1:
-    data = formatAnvlRequest(args[1:])
-  else:
-    data = None
-  response = issueRequest("id/"+encode(id), "POST", data)
-  printAnvlResponse(response)
-elif operation == "delete":
-  id = args[0]
-  response = issueRequest("id/"+encode(id), "DELETE")
-  printAnvlResponse(response)
-elif operation == "login":
-  response, headers = issueRequest("login", "GET", returnHeaders=True)
-  response += "\nsessionid=%s\n" %\
-    headers["set-cookie"].split(";")[0].split("=")[1]
-  printAnvlResponse(response)
-elif operation == "logout":
-  response = issueRequest("logout", "GET")
-  printAnvlResponse(response)
-elif operation == "status":
-  response = issueRequest("status", "GET")
-  printAnvlResponse(response)
+if __name__ == "__main__":
+  main()
